# Refactored Track to use boost::container::small_vector with optimised vector size