--- a/Framework/Geometry/src/Objects/Track.cpp
+++ b/Framework/Geometry/src/Objects/Track.cpp
@@ -9,6 +9,7 @@
 #include "MantidKernel/Matrix.h"
 #include "MantidKernel/Tolerance.h"
 #include "MantidKernel/V3D.h"
+#include <boost/iterator/distance.hpp>
 
 #include <algorithm>
 #include <cmath>
@@ -77,7 +78,7 @@ int Track::nonComplete() const {
 
   while (bc != m_links.end()) {
     if ((ac->exitPoint).distance(bc->entryPoint) > Tolerance) {
-      return (static_cast<int>(distance(m_links.begin(), bc)) + 1);
+      return (static_cast<int>(boost::distance(m_links.begin(), bc)) + 1);
     }
     ++ac;
     ++bc;