**Description of work.**
I Refactored Track to use `boost::container::small_vector` with an optimised vector size of 5.
The run times for the Montecarlo algorithms when running the script in #25628 for various vector sizes are:
N - Time
1            - 10 mins 55.01 secs
4            - 10 mins 47.06 secs
**5            - 10 mins 30.44 secs**
6            - 10 mins 38.16 secs
7            - 10 mins 41.17 secs
10          - 10 mins 41.29 secs
100        - 10 mins 35.27 secs

Using `std::vector`, the same script takes 11 minutes 7.76 seconds.
<!-- If the original issue was raised by a user they should be named here. Do not leak email addresses
**Report to:** [user name]
-->

<!-- Instructions for testing. -->

Fixes #25642 . <!-- and fix #xxxx or close #xxxx xor resolves #xxxx -->
<!-- alternative
*There is no associated issue.*
-->

<!-- delete this if you added release notes
*This does not require release notes* because **fill in an explanation of why**
-->

---

#### Reviewer ####

Please comment on the following ([full description](http://developer.mantidproject.org/IndividualTicketTesting/)):

##### Code Review #####

- Is the code of an acceptable quality?
- Does the code conform to the [coding standards](http://developer.mantidproject.org/Standards/)?
- Are the unit tests small and test the class in isolation?
- If there are changes in the release notes then do they describe the changes appropriately?

##### Functional Tests #####

- Do changes function as described? Add comments below that describe the tests performed?
- Do the changes handle unexpected situations, e.g. bad input?
- Has the relevant (user and developer) documentation been added/updated?

Does everything look good? Mark the review as **Approve**. A member of `@mantidproject/gatekeepers` will take care of it.