@peterfpeterson 2019-05-03
Please add timing for using `std::vector` as a reference point.