@aybamidele 2019-05-07
I've edited the comment to include the timing for `std::vector` obtained in #25628