@martyngigg 2019-05-08
This looks good but our first problem is that `small_vector` only appeared in boost 1.58 and RHEL currently has 1.53. Apparently boost 1.69 is available in EPEL though. I think @DanNixon required a newer version for a sort algorithm?