@DanNixon 2019-05-08
@martyngigg Previously yes, but I did some benchmarks on real data and the sort algorithm I wanted from Boost was only marginally better than TBB's `parallel_sort`.

On another note, are there performance tests that cover the improvements here? If not it may be worth adding one first.