@mantid-builder 2019-05-22
@aybamidele have you had a chance to see why the build is failing?
<!-- {"problem_type": "failing"} -->