@mantid-builder 2019-06-19
@aybamidele the build is failing, can you investigate.
<!-- {"problem_type": "failing"} -->