@mantid-builder 2019-08-15
@aybamidele there are conflicts here, can you resolve them.
<!-- {"problem_type": "conflicting"} -->