@mantid-builder 2019-08-15
@aybamidele the build is failing, can you investigate.
<!-- {"problem_type": "failing"} -->