@mantid-builder 2019-09-28
@aybamidele there are conflicts here, can you resolve them.
<!-- {"problem_type": "conflicting"} -->