@mantid-builder 2019-10-13
@aybamidele there are conflicts here, can you resolve them.
<!-- {"problem_type": "conflicting"} -->