@mantid-builder 2019-11-23
@aybamidele the build is failing, can you investigate.
<!-- {"problem_type": "failing"} -->